+++ src/pages/Message/MessageWritePage.jsx
@@ -1,11 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import Form from '../../components/common/Form';
-import Mheader from '../../components/message/Mheader';
-import MListBox from '../../components/message/MListBox';
-import MRead from '../../components/message/MRead';
 import MWrite from '../../components/message/MWrite';
-import NavBar from '../../components/common/NavBar';
+import FormHeader from '../../components/common/FormHeader';
 
 const s = {
   Frame: styled.div`
@@ -24,7 +20,7 @@ const MessageWritePage = () => {
   return (
     <>
       <s.Frame>
-        <Mheader text="쪽지보내기" />
+        <FormHeader text="쪽지보내기" type="send" />
         <MWrite />
       </s.Frame>
     </>
